feat: introduce nuxt 3 simple app
BREAKING CHANGE: move app to nuxt 3
