+++ .npmrc
@@ -1,0 +1,1 @@
+shamefully-hoist=true
