--- config/config.js
@@ -1,427 +1,0 @@
-export const verbs = [
-    {
-        base: 'be',
-        past: ['was', 'were', 'was/were'],
-        participle: 'been',
-    },
-    {
-        base: 'become',
-        past: 'became',
-        participle: 'become',
-    },
-    {
-        base: 'begin',
-        past: 'began',
-        participle: 'begun',
-    },
-    {
-        base: 'bite',
-        past: 'bit',
-        participle: 'bitten',
-    },
-    {
-        base: 'blow',
-        past: 'blew',
-        participle: ['blown', 'blowed'],
-    },
-    {
-        base: 'break',
-        past: 'broke',
-        participle: 'broken',
-    },
-    {
-        base: 'bring',
-        past: 'brought',
-        participle: 'brought',
-    },
-    {
-        base: 'build',
-        past: 'built',
-        participle: 'built',
-    },
-    {
-        base: 'burn',
-        past: ['burnt', 'burned'],
-        participle: ['burnt', 'burned'],
-    },
-    {
-        base: 'buy',
-        past: 'bought',
-        participle: 'bought',
-    },
-    {
-        base: 'catch',
-        past: 'caught',
-        participle: 'caught',
-    },
-    {
-        base: 'choose',
-        past: 'chose',
-        participle: 'chosen',
-    },
-    {
-        base: 'come',
-        past: 'came',
-        participle: 'come',
-    },
-    {
-        base: 'cost',
-        past: 'cost',
-        participle: 'cost',
-    },
-    {
-        base: 'cut',
-        past: 'cut',
-        participle: 'cut',
-    },
-    {
-        base: 'do',
-        past: 'did',
-        participle: 'done',
-    },
-    {
-        base: 'draw',
-        past: 'drew',
-        participle: 'drawn',
-    },
-    {
-        base: 'drink',
-        past: 'drank',
-        participle: 'drunk',
-    },
-    {
-        base: 'drive',
-        past: 'drove',
-        participle: 'driven',
-    },
-    {
-        base: 'eat',
-        past: 'ate',
-        participle: 'eaten',
-    },
-    {
-        base: 'fall',
-        past: 'fell',
-        participle: 'fallen',
-    },
-    {
-        base: 'feed',
-        past: 'fed',
-        participle: 'fed',
-    },
-    {
-        base: 'feel',
-        past: 'felt',
-        participle: 'felt',
-    },
-    {
-        base: 'fight',
-        past: 'fought',
-        participle: 'fought',
-    },
-    {
-        base: 'find',
-        past: 'found',
-        participle: 'found',
-    },
-    {
-        base: 'fly',
-        past: 'flew',
-        participle: 'flown',
-    },
-    {
-        base: 'forget',
-        past: 'forgot',
-        participle: 'forgotten',
-    },
-    {
-        base: 'forgive',
-        past: 'forgave',
-        participle: 'forgiven',
-    },
-    {
-        base: 'get',
-        past: 'got',
-        participle: 'got',
-    },
-    {
-        base: 'give',
-        past: 'gave',
-        participle: 'given',
-    },
-    {
-        base: 'go',
-        past: 'went',
-        participle: 'gone',
-    },
-    {
-        base: 'grow',
-        past: 'grew',
-        participle: 'grown',
-    },
-    {
-        base: 'hang',
-        past: 'hung',
-        participle: 'hung',
-    },
-    {
-        base: 'have',
-        past: 'had',
-        participle: 'had',
-    },
-    {
-        base: 'hear',
-        past: 'heard',
-        participle: 'heard',
-    },
-    {
-        base: 'hide',
-        past: 'hid',
-        participle: 'hidden',
-    },
-    {
-        base: 'hit',
-        past: 'hit',
-        participle: 'hit',
-    },
-    {
-        base: 'hold',
-        past: 'held',
-        participle: 'held',
-    },
-    {
-        base: 'hurt',
-        past: 'hurt',
-        participle: 'hurt',
-    },
-    {
-        base: 'keep',
-        past: 'kept',
-        participle: 'kept',
-    },
-    {
-        base: 'know',
-        past: 'knew',
-        past: 'known',
-    },
-    {
-        base: 'learn',
-        past: ['lernt', 'learned'],
-        participle: ['learnt', 'learned'],
-    },
-    {
-        base: 'leave',
-        past: 'left',
-        participle: 'left',
-    },
-    {
-        base: 'let',
-        past: 'let',
-        participle: 'let',
-    },
-    {
-        base: 'lose',
-        past: 'lost',
-        participle: 'lost',
-    },
-    {
-        base: 'make',
-        past: 'made',
-        participle: 'made',
-    },
-    {
-        base: 'mean',
-        past: 'meant',
-        participle: 'meant',
-    },
-    {
-        base: 'meet',
-        past: 'met',
-        participle: 'met',
-    },
-    {
-        base: 'pay',
-        past: 'paid',
-        participle: 'paid',
-    },
-    {
-        base: 'put',
-        past: 'put',
-        participle: 'put',
-    },
-    {
-        base: 'read',
-        past: 'read',
-        participle: 'read',
-    },
-    {
-        base: 'ride',
-        past: 'rode',
-        participle: 'ridden',
-    },
-    {
-        base: 'ring',
-        past: 'rang',
-        participle: 'rung',
-    },
-    {
-        base: 'run',
-        past: 'ran',
-        participle: 'run',
-    },
-    {
-        base: 'say',
-        past: 'said',
-        participle: 'said',
-    },
-    {
-        base: 'see',
-        past: 'saw',
-        participle: 'seen',
-    },
-    {
-        base: 'sell',
-        past: 'sold',
-        participle: 'sold',
-    },
-    {
-        base: 'send',
-        past: 'sent',
-        participle: 'sent',
-    },
-    {
-        base: 'set',
-        past: 'set',
-        participle: 'set',
-    },
-    {
-        base: 'shine',
-        past: 'shone',
-        participle: 'shone',
-    },
-    {
-        base: 'show',
-        past: 'showed',
-        participle: ['shown', 'showed'],
-    },
-    {
-        base: 'shut',
-        past: 'shut',
-        participle: 'shut',
-    },
-    {
-        base: 'sing',
-        past: 'sang',
-        participle: 'sung',
-    },
-    {
-        base: 'sit',
-        past: 'sat',
-        participle: 'sat',
-    },
-    {
-        base: 'sleep',
-        past: 'slept',
-        participle: 'slept',
-    },
-    {
-        base: 'smell',
-        past: 'slept',
-        participle: 'slept',
-    },
-    {
-        base: 'smell',
-        past: ['smelt', 'smelled'],
-        participle: ['smelt', 'smelled'],
-    },
-    {
-        base: 'speak',
-        past: 'spoke',
-        participle: 'spoken',
-    },
-    {
-        base: 'spell',
-        past: ['spelt', 'spelled'],
-        participle: ['spelt', 'spelled'],
-    },
-    {
-        base: 'spend',
-        past: 'spent',
-        participle: 'spent',
-    },
-    {
-        base: 'spin',
-        past: 'spun',
-        participle: 'spun',
-    },
-    {
-        base: 'spread',
-        past: 'spread',
-        participle: 'spread',
-    },
-    {
-        base: 'stand',
-        past: 'stood',
-        participle: 'stood',
-    },
-    {
-        base: 'steal',
-        past: 'stole',
-        participle: 'stolen',
-    },
-    {
-        base: 'swim',
-        past: 'swam',
-        participle: 'swum',
-    },
-    {
-        base: 'take',
-        past: 'took',
-        participle: 'taken',
-    },
-    {
-        base: 'teach',
-        past: 'taught',
-        participle: 'taught',
-    },
-    {
-        base: 'tell',
-        past: 'told',
-        participle: 'told',
-    },
-    {
-        base: 'think',
-        past: 'thought',
-        participle: 'thought',
-    },
-    {
-        base: 'throw',
-        past: 'threw',
-        participle: 'thrown',
-    },
-    {
-        base: 'understand',
-        past: 'understood',
-        participle: 'understood',
-    },
-    {
-        base: 'wake',
-        past: 'woke',
-        participle: 'woken',
-    },
-    {
-        base: 'wear',
-        paste: 'wore',
-        participle: 'worn',
-    },
-    {
-        base: 'win',
-        paste: 'won',
-        participle: 'won',
-    },
-    {
-        base: 'write',
-        paste: 'wrote',
-        participle: 'written',
-    },
-];
