+++ config/verbs.js
@@ -1,0 +1,802 @@
+export const verbs = [
+  {
+    base: 'be',
+    past: ['was', 'were', 'was/were'],
+    participle: 'been',
+    translation: {
+      it: ['essere'],
+    },
+  },
+  {
+    base: 'become',
+    past: 'became',
+    participle: 'become',
+    translation: {
+      it: ['diventare'],
+    },
+  },
+  {
+    base: 'begin',
+    past: 'began',
+    participle: 'begun',
+    translation: {
+      it: ['iniziare'],
+    },
+  },
+  {
+    base: 'bite',
+    past: 'bit',
+    participle: 'bitten',
+    translation: {
+      it: ['mordere'],
+    },
+  },
+  {
+    base: 'blow',
+    past: 'blew',
+    participle: ['blown', 'blowed'],
+    translation: {
+      it: ['soffiare'],
+    },
+  },
+  {
+    base: 'break',
+    past: 'broke',
+    participle: 'broken',
+    translation: {
+      it: ['rompere'],
+    },
+  },
+  {
+    base: 'bring',
+    past: 'brought',
+    participle: 'brought',
+    translation: {
+      it: ['portare'],
+    },
+  },
+  {
+    base: 'build',
+    past: 'built',
+    participle: 'built',
+    translation: {
+      it: ['costruire'],
+    },
+  },
+  {
+    base: 'burn',
+    past: ['burnt', 'burned'],
+    participle: ['burnt', 'burned'],
+    translation: {
+      it: ['bruciare'],
+    },
+  },
+  {
+    base: 'buy',
+    past: 'bought',
+    participle: 'bought',
+    translation: {
+      it: ['comprare'],
+    },
+  },
+  {
+    base: 'catch',
+    past: 'caught',
+    participle: 'caught',
+    translation: {
+      it: ['prendere', 'afferrare'],
+    },
+  },
+  {
+    base: 'choose',
+    past: 'chose',
+    participle: 'chosen',
+    translation: {
+      it: ['scegliere'],
+    },
+  },
+  {
+    base: 'come',
+    past: 'came',
+    participle: 'come',
+    translation: {
+      it: ['venire'],
+    },
+  },
+  {
+    base: 'cost',
+    past: 'cost',
+    participle: 'cost',
+    translation: {
+      it: ['costare'],
+    },
+  },
+  {
+    base: 'cut',
+    past: 'cut',
+    participle: 'cut',
+    translation: {
+      it: ['tagliare'],
+    },
+  },
+  {
+    base: 'deal',
+    past: 'dealt',
+    participle: 'dealt',
+    translation: {
+      it: ['trattare'],
+    },
+  },
+  {
+    base: 'do',
+    past: 'did',
+    participle: 'done',
+    translation: {
+      it: ['fare'],
+    },
+  },
+  {
+    base: 'draw',
+    past: 'drew',
+    participle: 'drawn',
+    translation: {
+      it: ['disegnare'],
+    },
+  },
+  {
+    base: 'dream',
+    past: ['dreamt', 'dreamed'],
+    participle: ['dreamt', 'dreamed'],
+    translation: {
+      it: ['sognare'],
+    },
+  },
+  {
+    base: 'drink',
+    past: 'drank',
+    participle: 'drunk',
+    translation: {
+      it: ['bere'],
+    },
+  },
+  {
+    base: 'drive',
+    past: 'drove',
+    participle: 'driven',
+    translation: {
+      it: ['guidare'],
+    },
+  },
+  {
+    base: 'eat',
+    past: 'ate',
+    participle: 'eaten',
+    translation: {
+      it: ['mangiare'],
+    },
+  },
+  {
+    base: 'fall',
+    past: 'fell',
+    participle: 'fallen',
+    translation: {
+      it: ['cadere'],
+    },
+  },
+  {
+    base: 'feed',
+    past: 'fed',
+    participle: 'fed',
+    translation: {
+      it: ['nutrire'],
+    },
+  },
+  {
+    base: 'feel',
+    past: 'felt',
+    participle: 'felt',
+    translation: {
+      it: ['sentire'],
+    },
+  },
+  {
+    base: 'fight',
+    past: 'fought',
+    participle: 'fought',
+    translation: {
+      it: ['combattere'],
+    },
+  },
+  {
+    base: 'find',
+    past: 'found',
+    participle: 'found',
+    translation: {
+      it: ['trovare'],
+    },
+  },
+  {
+    base: 'fly',
+    past: 'flew',
+    participle: 'flown',
+    translation: {
+      it: ['volare'],
+    },
+  },
+  {
+    base: 'forget',
+    past: 'forgot',
+    participle: 'forgotten',
+    translation: {
+      it: ['dimenticare'],
+    },
+  },
+  {
+    base: 'forgive',
+    past: 'forgave',
+    participle: 'forgiven',
+    translation: {
+      it: ['perdonare'],
+    },
+  },
+  {
+    base: 'freeze',
+    past: 'froze',
+    participle: 'frozen',
+    translation: {
+      it: ['gelare'],
+    },
+  },
+  {
+    base: 'get',
+    past: 'got',
+    participle: 'got',
+    translation: {
+      it: ['ottenere'],
+    },
+  },
+  {
+    base: 'give',
+    past: 'gave',
+    participle: 'given',
+    translation: {
+      it: ['dare'],
+    },
+  },
+  {
+    base: 'go',
+    past: 'went',
+    participle: 'gone',
+    translation: {
+      it: ['andare'],
+    },
+  },
+  {
+    base: 'grow',
+    past: 'grew',
+    participle: 'grown',
+    translation: {
+      it: ['crescere'],
+    },
+  },
+  {
+    base: 'hang',
+    past: 'hung',
+    participle: 'hung',
+    translation: {
+      it: ['appendere'],
+    },
+  },
+  {
+    base: 'have',
+    past: 'had',
+    participle: 'had',
+    translation: {
+      it: ['avere'],
+    },
+  },
+  {
+    base: 'hear',
+    past: 'heard',
+    participle: 'heard',
+    translation: {
+      it: ['udire'],
+    },
+  },
+  {
+    base: 'hide',
+    past: 'hid',
+    participle: 'hidden',
+    translation: {
+      it: ['nascondere'],
+    },
+  },
+  {
+    base: 'hit',
+    past: 'hit',
+    participle: 'hit',
+    translation: {
+      it: ['colpire'],
+    },
+  },
+  {
+    base: 'hold',
+    past: 'held',
+    participle: 'held',
+    translation: {
+      it: ['tenere'],
+    },
+  },
+  {
+    base: 'hurt',
+    past: 'hurt',
+    participle: 'hurt',
+    translation: {
+      it: ['ferire'],
+    },
+  },
+  {
+    base: 'keep',
+    past: 'kept',
+    participle: 'kept',
+    translation: {
+      it: ['tenere'],
+    },
+  },
+  {
+    base: 'know',
+    past: 'knew',
+    past: 'known',
+    translation: {
+      it: ['conoscere', 'sapere'],
+    },
+  },
+  {
+    base: 'lay',
+    past: 'laid',
+    participle: 'laid',
+    translation: {
+      it: ['giacere'],
+    },
+  },
+  {
+    base: 'lead',
+    past: 'led',
+    participle: 'led',
+    translation: {
+      it: ['condurre'],
+    },
+  },
+  {
+    base: 'learn',
+    past: ['lernt', 'learned'],
+    participle: ['learnt', 'learned'],
+    translation: {
+      it: ['imparare'],
+    },
+  },
+  {
+    base: 'leave',
+    past: 'left',
+    participle: 'left',
+    translation: {
+      it: ['lasciare', 'partire'],
+    },
+  },
+  {
+    base: 'lend',
+    past: 'lent',
+    participle: 'lent',
+    translation: {
+      it: ['prestare'],
+    },
+  },
+  {
+    base: 'let',
+    past: 'let',
+    participle: 'let',
+    translation: {
+      it: ['lasciare', 'permettere'],
+    },
+  },
+  {
+    base: 'light',
+    past: 'lit',
+    participle: 'lit',
+    translation: {
+      it: ['accendere', 'illuminare'],
+    },
+  },
+  {
+    base: 'lose',
+    past: 'lost',
+    participle: 'lost',
+    translation: {
+      it: ['perdere'],
+    },
+  },
+  {
+    base: 'make',
+    past: 'made',
+    participle: 'made',
+    translation: {
+      it: ['fare'],
+    },
+  },
+  {
+    base: 'mean',
+    past: 'meant',
+    participle: 'meant',
+    translation: {
+      it: ['significare', 'voler dire'],
+    },
+  },
+  {
+    base: 'meet',
+    past: 'met',
+    participle: 'met',
+    translation: {
+      it: ['incontrare'],
+    },
+  },
+  {
+    base: 'pay',
+    past: 'paid',
+    participle: 'paid',
+    translation: {
+      it: ['pagare'],
+    },
+  },
+  {
+    base: 'put',
+    past: 'put',
+    participle: 'put',
+    translation: {
+      it: ['mettere'],
+    },
+  },
+  {
+    base: 'read',
+    past: 'read',
+    participle: 'read',
+    translation: {
+      it: ['leggere'],
+    },
+  },
+  {
+    base: 'ride',
+    past: 'rode',
+    participle: 'ridden',
+    translation: {
+      it: ['cavalcare'],
+    },
+  },
+  {
+    base: 'ring',
+    past: 'rang',
+    participle: 'rung',
+    translation: {
+      it: ['suonare'],
+    },
+  },
+  {
+    base: 'rise',
+    past: 'rose',
+    participle: 'risen',
+    translation: {
+      it: ['sorgere', 'alzarsi'],
+    },
+  },
+  {
+    base: 'run',
+    past: 'ran',
+    participle: 'run',
+    translation: {
+      it: ['correre'],
+    },
+  },
+  {
+    base: 'say',
+    past: 'said',
+    participle: 'said',
+    translation: {
+      it: ['dire'],
+    },
+  },
+  {
+    base: 'see',
+    past: 'saw',
+    participle: 'seen',
+    translation: {
+      it: ['vedere'],
+    },
+  },
+  {
+    base: 'sell',
+    past: 'sold',
+    participle: 'sold',
+    translation: {
+      it: ['vendere'],
+    },
+  },
+  {
+    base: 'send',
+    past: 'sent',
+    participle: 'sent',
+    translation: {
+      it: ['spedire', 'mandare'],
+    },
+  },
+  {
+    base: 'set',
+    past: 'set',
+    participle: 'set',
+    translation: {
+      it: ['impostare'],
+    },
+  },
+  {
+    base: 'shake',
+    past: 'shook',
+    participle: 'shaken',
+    translation: {
+      it: ['scuotere'],
+    },
+  },
+  {
+    base: 'shine',
+    past: 'shone',
+    participle: 'shone',
+    translation: {
+      it: ['brillare', 'splendere'],
+    },
+  },
+  {
+    base: 'shoot',
+    past: 'shot',
+    participle: 'shot',
+    translation: {
+      it: ['sparare'],
+    },
+  },
+  {
+    base: 'show',
+    past: 'showed',
+    participle: ['shown', 'showed'],
+    translation: {
+      it: ['mostrare', 'far vedere'],
+    },
+  },
+  {
+    base: 'shut',
+    past: 'shut',
+    participle: 'shut',
+    translation: {
+      it: ['chiudere'],
+    },
+  },
+  {
+    base: 'sing',
+    past: 'sang',
+    participle: 'sung',
+    translation: {
+      it: ['cantare'],
+    },
+  },
+  {
+    base: 'sink',
+    past: 'sank',
+    participle: 'sunk',
+    translation: {
+      it: ['affondare'],
+    },
+  },
+  {
+    base: 'sit',
+    past: 'sat',
+    participle: 'sat',
+    translation: {
+      it: ['sedere'],
+    },
+  },
+  {
+    base: 'sleep',
+    past: 'slept',
+    participle: 'slept',
+    translation: {
+      it: ['dormire'],
+    },
+  },
+  {
+    base: 'smell',
+    past: ['smelt', 'smelled'],
+    participle: ['smelt', 'smelled'],
+    translation: {
+      it: ['annusare'],
+    },
+  },
+  {
+    base: 'speak',
+    past: 'spoke',
+    participle: 'spoken',
+    translation: {
+      it: ['parlare'],
+    },
+  },
+  {
+    base: 'spell',
+    past: ['spelt', 'spelled'],
+    participle: ['spelt', 'spelled'],
+    translation: {
+      it: ['scrivere (lettera per lettera)'],
+    },
+  },
+  {
+    base: 'spend',
+    past: 'spent',
+    participle: 'spent',
+    translation: {
+      it: ['trascorrere', 'spendere'],
+    },
+  },
+  {
+    base: 'spill',
+    past: ['spilt', 'spilled'],
+    participle: ['spilt', 'spilled'],
+    translation: {
+      it: ['versare'],
+    },
+  },
+  {
+    base: 'spin',
+    past: 'spun',
+    participle: 'span',
+    translation: {
+      it: ['girare'],
+    },
+  },
+  {
+    base: 'split',
+    past: 'split',
+    participle: 'split',
+    translation: {
+      it: ['dividere'],
+    },
+  },
+  {
+    base: 'spoil',
+    past: ['spoilt', 'spoiled'],
+    participle: ['spoilt', 'spoiled'],
+    translation: {
+      it: ['rovinare'],
+    },
+  },
+  {
+    base: 'spread',
+    past: 'spread',
+    participle: 'spread',
+    translation: {
+      it: ['spargere'],
+    },
+  },
+  {
+    base: 'stand',
+    past: 'stood',
+    participle: 'stood',
+    translation: {
+      it: ['stare in piedi'],
+    },
+  },
+  {
+    base: 'steal',
+    past: 'stole',
+    participle: 'stolen',
+    translation: {
+      it: ['rubare'],
+    },
+  },
+  {
+    base: 'stick',
+    past: 'stuck',
+    participle: 'stuck',
+    translation: {
+      it: ['attaccare'],
+    },
+  },
+  {
+    base: 'swim',
+    past: 'swam',
+    participle: 'swum',
+    translation: {
+      it: ['nuotare'],
+    },
+  },
+  {
+    base: 'take',
+    past: 'took',
+    participle: 'taken',
+    translation: {
+      it: ['prendere'],
+    },
+  },
+  {
+    base: 'teach',
+    past: 'taught',
+    participle: 'taught',
+    translation: {
+      it: ['insegnare'],
+    },
+  },
+  {
+    base: 'tell',
+    past: 'told',
+    participle: 'told',
+    translation: {
+      it: ['dire', 'raccontare'],
+    },
+  },
+  {
+    base: 'think',
+    past: 'thought',
+    participle: 'thought',
+    translation: {
+      it: ['pensare'],
+    },
+  },
+  {
+    base: 'throw',
+    past: 'threw',
+    participle: 'thrown',
+    translation: {
+      it: ['lanciare', 'buttare', 'gettare'],
+    },
+  },
+  {
+    base: 'understand',
+    past: 'understood',
+    participle: 'understood',
+    translation: {
+      it: ['capire'],
+    },
+  },
+  {
+    base: 'upset',
+    past: 'upset',
+    participle: 'upset',
+    translation: {
+      it: ['offendere'],
+    },
+  },
+  {
+    base: 'wake',
+    past: 'woke',
+    participle: 'woken',
+    translation: {
+      it: ['svegliare', 'svegliarsi'],
+    },
+  },
+  {
+    base: 'wear',
+    paste: 'wore',
+    participle: 'worn',
+    translation: {
+      it: ['indossare'],
+    },
+  },
+  {
+    base: 'win',
+    paste: 'won',
+    participle: 'won',
+    translation: {
+      it: ['vincere'],
+    },
+  },
+  {
+    base: 'write',
+    paste: 'wrote',
+    participle: 'written',
+    translation: {
+      it: ['scrivere'],
+    },
+  },
+];
