--- nuxt.config.js
@@ -1,56 +1,0 @@
-export default {
-    // Disable server-side rendering: https://go.nuxtjs.dev/ssr-mode
-    ssr: false,
-
-    // Target: https://go.nuxtjs.dev/config-target
-    target: 'static',
-
-    // Global page headers: https://go.nuxtjs.dev/config-head
-    head: {
-        title: 'irregular-verbs',
-        htmlAttrs: {
-            lang: 'en'
-        },
-        meta: [
-            { charset: 'utf-8' },
-            { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-            { hid: 'description', name: 'description', content: '' },
-            { name: 'format-detection', content: 'telephone=no' }
-        ],
-        link: [
-            { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }
-        ]
-    },
-
-    loadingIndicator: {
-        name: 'circle',
-        color: '#3B8070',
-        background: 'white'
-    },
-
-    // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
-    plugins: [
-    ],
-
-    // Auto import components: https://go.nuxtjs.dev/config-components
-    components: true,
-
-    // Modules for dev and build (recommended): https://go.nuxtjs.dev/config-modules
-    buildModules: [
-        '@nuxtjs/tailwindcss',
-    ],
-
-    // Modules: https://go.nuxtjs.dev/config-modules
-    modules: [
-    ],
-
-    // Build Configuration: https://go.nuxtjs.dev/config-build
-    build: {
-        extend(config, { isClient }) {
-            // Extend only webpack config for client-bundle
-            if (isClient) {
-                config.devtool = 'source-map'
-            }
-        }
-    }
-}
