--- tailwind.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-  content: [
-    './components/**/*.{vue,js}',
-    './layouts/**/*.vue',
-    './pages/**/*.vue',
-    './plugins/**/*.{js,ts}',
-    './nuxt.config.{js,ts}',
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
